Just an idea: using a thread for a specific interval
and not a thread for every single module. Probably conservers more
energy as it is more efficient on modern CPUs to do a bunch of work
in one piece as opposed to doing it in small pieces.

Would also avoid some overhead with the GIL.

The problem, on the other hand is how one copes with slow, blocking modules
like something bound to internet access.
Basically two ideas come to my mind:
1) Introduce another module type (bad!), as shown here, for this,
which is not pooled.
2) Measure the execution time of every module and if a run takes longer than
the interval (sleep period) move the time consuming tasks into their own
threads until they get fast enough to get moved back into the pool.
I kinda like the idea of that self-organizing thread pool... and it wouldn't
change the API, which I would really like to stay frozen ATM.