Introduce a colorize option for the weather status plugin.
Add an option to change the color and add an icon in the display based
on the text returned from weather.com. For example, if the waether is
considered "Sunny" by weather.com the weather text will be set to a
golden yellow and a little snow will be added to the display.