Try/Except the pywapi call
Often, the pywapi.get_weather... call will yield a "'NoneType' object has no attribute 'data' error" if the weather call does not work properly. 
This is usually resolved on the next call, but leads to an ugly warning in the status bar.
I simply wrapped the pywapi call in a try-except block where the except condition returns False. (Perhaps instead it should output something to a log file?)
This solved the problem for me.

PS: This is my first attempt at proposing a change, so if I'm not doing it right I'd love to be pointed to an intro to contributing to Github projects or something similar.