net_speed: a module to display internet speeds
This module provides a glimpse/snapshot of current internet based
speeds.  To keep things simple (and due to bugs in speedtest_cli), this
module simply uses requests to download a small(ish) image and times the
download.  This is not a perfect/definitive test, but it should be a
decent indicator of how the internet bound traffic looks on the clients
network.