Add Syncthing module
This module provides support for Syncthing [1]. Currently this module
only supports showing the up/down status of Syncthing and it is possible
to start/shutdown Syncthing via click events. A few callback functions
for usage with/without systemd are provided as well.

The module is designed in a generic way (st_get(), st_post()), such that
new features could be add very easily.

[1]: https://syncthing.net