Add wunderground module
This module tries to use as much of the same variable naming conventions
that the ``weather`` module uses as possible, to make transitioning
easier in the future in case we decide to make a base class for all
modules which provide weather data.

An API key is required to use this module, information on obtaining one
can be found in the docstring.