clock: Detect system TZ (if possible) for cases where no TZ is specified
When the pytz dep was reintroduced, this made all TZ usage reliant upon
a tzinfo object generated by pytz. This had the effect of making the
``%Z`` strftime placeholder evaluate as blank when a timezone is not
explicitly defined (i.e. when the format is just a format string and
not a tuple containing the format string and timezone).

This commit resolves this regression by using pytz to build a tzinfo
object from either /etc/localtime or /etc/timezone during the ``init()``
function, establishing the system TZ to be used for those cases where no
TZ was specified.