Fix traceback in weather.com weather backend
It seems that this XML feed may have been changed, as it now returns an
error saying that the location is not valid.

This fixes the traceback when we try to run str functions on None when
the XML keys don't exist in the return data. However, the long term fix
will be to either update the backend to use the updated formatting
(which may be hard as this XML feed appears to be undocumented), or to
simply retire this backend.