uptime: Check seconds_alert against raw seconds
Originally, uptime checked seconds_alert against seconds to determine
whether to use color_alert, but if {days}, {hours}, or {mins} is used in
the format string, seconds is truncated to exclude anything above a day,
hour, or minute respectively.
This caused seconds to always be below a seconds_alert value greater
than a day, for example, if {days} is used, so the alert color was never
used.
Fix this by saving and checking against the raw seconds that were
originally read.