pulseaudio: remove output when sink is gone
When `sink` is set and you have USB card or Bluetooth headset.

When the sink is created, the output is set and the module appear.

But when the sink is deleted, the output is not deleted, and the module
stay.

This change detects when the sink is deleted and delete the output.